iotedge/EdgeSolution/modules/AnomalyDetectionModule/requirements.txt
azure-iot-device~=2.7.0
pandas
joblib
scikit-learn